[FIX] opcion de no traerse la bd de prod

+++ restore_database.sh
@@ -2,7 +2,7 @@
 #######################################################################
 # Restore newest backup from production to local deactivating database.
 # parameters: client version
-# por ejemplo ./restore_database.sh polimera 11.0
+# por ejemplo ./restore_database.sh polimera 11.0e
 
 client=$1
 version=$2
@@ -26,13 +26,17 @@ server=$username"@"$client
 dir="/odoo_ar/odoo-"$version/$client
 bkp=$dir"/backup_dir"
 
-# eliminar los backups locales
-echo "Remove local backups in "$bkp
-sudo rm -r $bkp/*
+# si hay algun parametro en 3 NO traigo la bd de produccion
+if [ -z "$3" ]
+then
+   # eliminar los backups locales
+   echo "Remove local backups in "$bkp
+   sudo rm -r $bkp/*
 
-# traerse el backup mas nuevo
-echo "Getting newest backup from "$server
-scp $server:$bkp/$(ssh $server "ls -t $bkp/ | head -1") $bkp/
+   # traerse el backup mas nuevo
+   echo "Getting newest backup from "$server
+   scp $server:$bkp/$(ssh $server "ls -t $bkp/ | head -1") $bkp/
+fi
 
 # restorear el backup mas nuevo que encuentre
 sudo docker run --rm -i \
